% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustLeftMargin.R
\name{adjustLeftMargin}
\alias{adjustLeftMargin}
\title{Adjust left margin of ggplot2 object to specific margin}
\usage{
adjustLeftMargin(plotdat, chars_n = 10)
}
\arguments{
\item{plotdat}{ggplot2 object}

\item{nchar4ylabels}{integer. Number of "A" character y values. Will result in changing width for left margin.}
}
\description{
Function creates dummy plot with specific width for left margin, by creating arbitrary y values: Repetition of character "A". Number of repetion can be specified.
The resulting width for left in dummy plot will be extracted using ggtable_build().
The grobs of target plot object will be adjustet using ggtable_build(). The width for left margin of the dummy plot will overwrite the width for left margin of the target plot.
}
