% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakStringToLines.R
\name{breakStringToLines}
\alias{breakStringToLines}
\title{Break text into new lines by max. characters per laine}
\usage{
breakStringToLines(text, max_nchar = NULL, max_nlines = NULL,
  max_nwords = NULL)
}
\description{
Break text into new lines by max. characters per laine
}
