% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.kmeanspp.R
\name{clust.kmeanspp}
\alias{clust.kmeanspp}
\title{k-Means++ clustering}
\usage{
clust.kmeanspp(dat_wide, k = 3, start = "random", iter.max = 500,
  nstart = 1, algorithm = "Hartigan-Wong", ...)
}
\description{
k-Means++ clustering
}
