% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelOutlierGrubb.R
\name{codeOutlierGrubb}
\alias{codeOutlierGrubb}
\title{Flag outlier according to Grubb (1969)}
\usage{
codeOutlierGrubb(dat)
}
\description{
http://daryavanichkina.com/wikipages/r/outlierdetectionR.html
}

