% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeScores.R
\name{computeScores}
\alias{computeScores}
\title{Recode/Invert items}
\usage{
computeScores(dat, name4score, list4items, fun = "mean", compZ = F)
}
\arguments{
\item{dat}{Data object}

\item{list4items}{List of item/column names for each score in name4score}

\item{fun}{Set function for score computation as character: "mean" or "sum"}

\item{name4scores}{Names for target scores}
}
\description{
Recodes items from questionnaires
}

