% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummary.R
\name{computeSummary}
\alias{computeSummary}
\title{Compute summary}
\usage{
computeSummary(dat, col_names_group = NULL, col_names_values = "value",
  fun_name_stat = c("mean", "sd", "median", "min", "max", "se", "ci"),
  conf = 0.95, remove_na = T)
}
\description{
Compute summary
}
