% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.GPS2XY.R
\name{conv.GPS2XY}
\alias{conv.GPS2XY}
\title{Convert GPS to XY (m)}
\usage{
conv.GPS2XY(dat, col_name_gps_lon, col_name_gps_lat, gps_lon_origin,
  gps_lat_origin, return_only_x = F, return_only_y = F)
}
\description{
Convert GPS to XY (m)
}
