% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.timestamp2sec.R
\name{conv.timestamp2sec}
\alias{conv.timestamp2sec}
\title{Convert timestamp (HH:MM:SS.ms) to seconds}
\usage{
conv.timestamp2sec(timestamp)
}
\arguments{
\item{timestamp}{character. Timestamp in format HH:MM:SS.s}
}
\description{
Convert timestamp (HH:MM:SS.ms) to seconds
}
