% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFilePath.R
\name{createFilePath}
\alias{createFilePath}
\title{Build file path and create subdirectories if necessary}
\usage{
createFilePath(dir_list)
}
\description{
Build file path and create subdirectories if necessary
}
