% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnectBySettings.R
\name{dbConnectBySettings}
\alias{dbConnectBySettings}
\title{Connect to PostgreSQL database}
\usage{
dbConnectBySettings(settings, dbname = settings$dns,
  dbconn_name = settings$dns, pwd)
}
\description{
Connect to PostgreSQL database
}

