% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFindConnObj.R
\name{dbFindConnObj}
\alias{dbFindConnObj}
\title{Find object name by dns}
\usage{
dbFindConnObj(db_name, output = F, ...)
}
\description{
Find object names by dns, if connection was initialised using dbConnectOperator, which adds the original dns as attribute to the object name of the connection.
}
