% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetSrc.R
\name{dbGetSrc}
\alias{dbGetSrc}
\title{Query complete database source}
\usage{
dbGetSrc(dbconn, name4src)
}
\arguments{
\item{dbconn}{A DBIConnection object, as produced by dbConnect or character of the object name}

\item{name4src}{Name of database source as character}
}
\description{
Sources my be tables or views
}
