% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachPackage.R
\name{detachPackage}
\alias{detachPackage}
\title{detachPackage}
\usage{
detachPackage(pkg_name, character.only = FALSE, force = FALSE)
}
\arguments{
\item{pkg_name}{Package name as object name or character}

\item{character.only}{logical. default = FALSE (function will deparse object name to character)}

\item{force}{logical. default = FALSE. If set to TRUE package will be detached even though other attached packages depend on it}
}
\description{
This function allows you unload packages quily by name (or as as string)
Source: http://stackoverflow.com/questions/6979917/how-to-unload-a-package-without-restarting-r
}
\examples{
#install.packages("ggplot2")
#library(ggplot2)
#unloadPackage(ggplot2)
#library(ggplot2)
#unloadPackage("ggplot2")
}
\keyword{detachPackage}

