% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunProc_status.R
\name{.outputFunProc_status}
\alias{.outputFunProc_status}
\title{Set outputFunProc active}
\usage{
.outputFunProc_status(status = F, print = F)
}
\arguments{
\item{status}{logical. default = T (active)}
}
\description{
Set outputFunProc active
}
