% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPercentageCurveGLM.R
\name{fitPercentageCurveGLM}
\alias{fitPercentageCurveGLM}
\title{Smooth values using GLM}
\usage{
fitPercentageCurveGLM(dat, col_name_reference, col_name_value,
  polypar = 4, family = quote(gaussian(link = "log")))
}
\description{
Smooth values using GLM
}
