% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDecimalPlaces.R
\name{getDecimalPlaces}
\alias{getDecimalPlaces}
\title{Get number of decimal places}
\usage{
getDecimalPlaces(values, output = puttytat4R_env$outputFunProc_status)
}
\description{
Necessary for correct value presentation after interpolation using spline
}

