% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initQuestSettings.R
\name{initQuestSettings}
\alias{initQuestSettings}
\title{Initialise settings for questionnaire}
\usage{
initQuestSettings(abbr, db_conn_name, db_src_name, sett_name = NULL,
  lang = "eng", max_nchar = 99, ...)
}
\description{
Initialise settings for questionnaire
}
