% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunCallStack.R
\name{outputFunCallStack}
\alias{outputFunCallStack}
\title{: Output function call stack}
\usage{
outputFunCallStack()
}
\description{
Outputs current frame number and objects, as well as frame number and objects from parents
}
