% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunProc.R
\name{outputFunProc}
\alias{outputFunProc}
\title{Print currently used function name and type}
\usage{
outputFunProc(proctype, funname = "", calling = T)
}
\arguments{
\item{proctype}{Single character for identifying process type: "I" for "Initialising", "R" for "Running", and "D" for "Done". No default. "I" and "D" can be used with any (arbitrary) function name without constrictions. "R" depends on active and parent calls.}

\item{funname}{String for function name. If no function name is given, outputFunProc will deparse function name from active call (as outputFunProc can be run inside another function).}
}
\description{
Print currently used function name and type
}

