% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputString.R
\name{outputString}
\alias{outputString}
\title{Print text with enclosing separating lines}
\usage{
outputString(string, type = "cat", sepline = F, seplinepos = c("T", "B"),
  seplinechar = "-", returnchar = F)
}
\arguments{
\item{string}{Vector of character values}

\item{position}{Vector of character values. Position of separating lines ("top", "bottom")}

\item{sepchar}{Character value for separating line}
}
\description{
Print multiple character values with enclosing separating lines
}

