% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputWSepLine.R
\name{outputWSepLine}
\alias{outputWSepLine}
\title{Print text with enclosing separating lines}
\usage{
outputWSepLine(output_txt, pos = c("top", "bottom"), sepchar = "-",
  returnchar = F)
}
\arguments{
\item{output_txt}{Vector of character values}

\item{sepchar}{Character value for separating line}

\item{position}{Vector of character values. Position of separating lines ("top", "bottom")}
}
\description{
Print multiple character values with enclosing separating lines
}

