% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotItemProfile.R
\name{plotItemProfile}
\alias{plotItemProfile}
\title{Plot item profile}
\usage{
plotItemProfile(dat, col_name_items = "variables",
  col_name_values = "mean", col_name_group = NULL, value_range)
}
\description{
Plot item profile
}
