% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressBar100.R
\docType{data}
\name{progressBar100}
\alias{progressBar100}
\title{Progressbar for 100 %}
\format{An object of class \code{txtProgressBar} of length 3.}
\usage{
progressBar100
}
\description{
Use txtProgressBar with predifined settings (0-100%)
}
\keyword{datasets}
