% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puttytat4R_env.R
\docType{data}
\name{puttytat4R_env}
\alias{puttytat4R_env}
\title{Test functions for outputFunProc}
\format{An object of class \code{environment} of length 1.}
\usage{
puttytat4R_env
}
\description{
Print multiple character values with enclosing separating lines
}
\keyword{datasets}

