% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshapeLong.R
\name{reshapeLong}
\alias{reshapeLong}
\title{Reshape data to long data frame}
\usage{
reshapeLong(dat, col_name_gathered_col_names = "col_name",
  col_name_gathered_values = "value", col_names_to_keep = NULL, ...)
}
\description{
Reshape data to long data frame
}
