% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeScores.R
\name{computeScores}
\alias{computeScores}
\title{Recode/Invert items}
\usage{
computeScores(dat, score_names, score_items_list, fun = "mean", compZ = F)
}
\arguments{
\item{dat}{Data object}

\item{fun}{Set function for score computation as character: "mean" or "sum"}

\item{score_namess}{Names for target scores}

\item{list4items}{List of item/column names for each score in score_names}
}
\description{
Recodes items from questionnaires
}
