% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeScores.R
\name{computeScores}
\alias{computeScores}
\title{Recode/Invert items}
\usage{
computeScores(dat, subscale_items_list, lower_subscale_names = T,
  fun = "mean", compute_z = F)
}
\arguments{
\item{dat}{Data object}

\item{fun}{Set function for score computation as character: "mean" or "sum"}

\item{subscale_namess}{Names for target scores}

\item{list4items}{List of item/column names for each score in subscale_names}
}
\description{
Recodes items from questionnaires
}
