% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnectBySettings.R
\name{dbConnectBySettings}
\alias{dbConnectBySettings}
\title{Connect to PostgreSQL database}
\usage{
dbConnectBySettings(settings, dbconn_name = "dbconn")
}
\description{
Connect to PostgreSQL database
}

