% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnectBySettings.R
\name{dbConnectBySettings}
\alias{dbConnectBySettings}
\title{Connect to PostgreSQL database}
\usage{
dbConnectBySettings(sett = sett_db_default, db_name = sett$dns,
  db_conn_name = sett$dns, pwd)
}
\description{
Connect to PostgreSQL database
}
