% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInitSettings.R
\name{dbInitSettings}
\alias{dbInitSettings}
\title{Initialise PostgreSQL settings}
\usage{
dbInitSettings(dir_name = "settings", file_name = "sett_db_default.R",
  dns_default = "\\"postgres\\"", set_to_default = F)
}
\description{
Create tempate for database settings in directory "settings"
}
