% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInitSettings.R
\name{dbInitSettings}
\alias{dbInitSettings}
\title{Initialise PostgreSQL settings}
\usage{
dbInitSettings(dirname = "db_set", filename = "db_set.R",
  dnsdefault = "postgres")
}
\description{
Create tempate for database settings in directory "settings"
}

