% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInitSettings.R
\name{dbInitSettings}
\alias{dbInitSettings}
\title{Initialise PostgreSQL settings}
\usage{
dbInitSettings(dir_name = "sett_db_default_default",
  filename = "sett_db_default_default.R", dns_default = "\\"postgres\\"")
}
\description{
Create tempate for database settings in directory "settings"
}
