% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInitSettings.R
\name{dbInitSettings}
\alias{dbInitSettings}
\title{Initialise PostgreSQL settings}
\usage{
dbInitSettings(dirname = "dbset", filename = "dbset.R")
}
\description{
Create tempate for database settings in directory "settings"
}

