% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeItems.R
\name{recodeItems}
\alias{recodeItems}
\title{Recode/Invert items}
\usage{
recodeItems(dat, name4cols, range)
}
\arguments{
\item{dat}{Data object}

\item{range}{Range of scale as integer}

\item{name4src}{Name of database source as character}
}
\description{
Recodes items from questionnaires
}
