% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reloadPackage.R
\name{reloadPackage}
\alias{reloadPackage}
\title{Print currently used function name and type}
\usage{
reloadPackage(name4pkg, character.only = FALSE)
}
\arguments{
\item{character.only}{logical. default = FALSE (function will deparse object name to character)}

\item{pkg_name}{Package name as object name or character}
}
\description{
Print currently used function name and type
}
